Move wikipedia.php to autoloadable utility class
